Move data in separate file. Break skills into individual sections.
